SWARM-365: Added tests for the Generate mojo
Motivation:
This test reproduces the errors caused when you update to Keycloak 1.9.1.Final, as requested in SWARM-321.
It makes easier to debug and fix the issue.

Result:
The code that reproduces SWARM-365  is commented out so the build won't fail. Uncomment it and run the test to reproduce the issue.