Have fraction-list task use exposed-components
Motivation:

The fraction-list task uses the calling project's dependencyManagement
to determine fractions. We only use it in one place:
wildfly-swarm/fraction-list/, which imports the BOM to populate
dependencyManagement. The BOM is provided by wildfly-swarm/bom/, but
maven processes dependencyManagement imports at model build
time, *before* the bom/ module builds, so fraction-list/ uses the BOM
from the *last* build. If we use the BOM-exposed components instead of
dependencyManagement, we don't have this chicken-and-egg problem.

Modifications:

* have FractionListMojo inherit from AbstractExposedComponentsMojo in
  order to use the exposed-components

Result:

A chicken without an egg!