Generate provided-dependencies.txt if one does not exist.
If the pom.xml contains one or more <dependency> element
with swarm:scope="provided" attribute, and a provided-dependencies.txt
doesn't exist, use those marked <dependency> items as the contents
of provided-dependenciex.txt.

The swarm:scope attribute is tied to the xmlns of xmlns:swarm="http://wildfly-swarm.io/"
which should be added to the root of the pom.xml at the <project> element.