Use the remote repos that maven gives us
The -swarm PR validation job fails to find an aether
class (org/eclipse/aether/util/repository/AuthenticationBuilder) when
running the BomMojo. It's unclear why this can't be found (it may be
related to the lack of a requiresDependencyResolution scope in the @Mojo
annotation, but that's just a guess). Anyway, it turns out we don't even
need the class, since we were using it to convert from maven
ArtifactRepositories into aether RemoteRepositories, when maven can give
us RemoteRepositories directly.