SWARM-1023: Arquillian should always be in BOM
Motivation
----------
Currently the Arquillian fraction is only added to a BOM based on its stability. However, we really want Arquillian to be present in every BOM for testing.

This also allows us to decouple the Arquillian build from fraction-list.

Modifications
-------------
Exclude Arquillian from list of possible fractions to include in a BOM, and then add it directly without resolving it in.

Result
------
Arquillian is now in all BOM irrespective of its stability.