SWARM-1282 - Support "certified" community BOM in product
Motivation:

We don't want to productize everything, but we do want to "certify"
some community bits against the productized bits.

Changes:

Add a generate-certified-bom mojo for generating a certified BOM.

It consumes (via a plugin dependency) a particular fraction-metadata,
along with a `certified.conf` in PWD listing the artifactId of the
items we desire to denote as certified.  In produces a :bom-certified
artifact which includes those things, plus our :arquillian fraction,
using the community version specified via the abovementioned
plugin dependency.

Result:

If used, another BOM.