SWARM-732: Produce Maven Repo for BOMs
Motivation
----------
Provide a zip artifact of Maven repo for a BOM that can be used for "air gap" situations.

Modifications
-------------
Add new Mojos to create generator project from a BOM and then build it with artifacts being installed to empty repo that we can zip.

Result
------
Add Mojo for creating a zipped Maven local repo and attach to Maven lifecycle.