SWARM-1467: "runtime" modules should be marked private
Motivation
----------
For each fraction, the `wildfly-swarm-fraction-plugin` generates
[up to] 4 modules (`module.xml`): `main`, `api`, `deployment` and
`runtime`. At the very least, the `runtime` one isn't supposed to
be used directly and compatibility isn't guaranteed. Those modules
should be marked _private_.

Modifications
-------------
Add the `jboss.api` property with value of `private` when generating
the `module.xml` for the `runtime` modules of all fractions.

Result
------
When user imports the `runtime` module of a fraction, they get
a friendly warning.