THORN-2163: don't use fraction-list.txt to generate bom-certified
Motivation
----------
For generating the certified BOM, the `thorntail-fraction-plugin`
reads `certified.conf`, cross-checks it with `fraction-list.txt`
and generates a BOM including only those fractions that are in
both lists.

When building product, the intent is that `boms/bom-certified`
brings community `io.thorntail:fraction-metadata` as a dependency
of `thorntail-fraction-plugin`, so the `fraction-list.txt` always
comes from community. This used to be true, but it no longer is:
product 2.2.0.Final will be based on community 2.2.0.Final, so
the `io.thorntail:fraction-metadata` artifact no longer comes
from community; it's produced by the same build that produces
the certified BOM. Hence we can't use the `fraction-list.txt`
anymore -- it doesn't and can't contain the community fractions
that should be present in the certified BOM.

Modifications
-------------
Stopped reading `fraction-list.txt` when generating certified BOM,
instead generate it based solely on `certified.conf`.

This requires passing the expected community version as a parameter
to the `thorntail-fraction-plugin` invocation, which is easy to do.

Result
------
Allows successfully generating certified BOM for product again.